requirements.txt
python-dotenv
easydict
celery
redis
requests
selenium
fake_useragent